** IMPORTANT **
For Xmas, here is a brand new API (that breaks all existing code).

We remove the starpu_data_interface_t enum type as it does not permit
to have the application to manipulate user-provided data interfaces.  This also
reduces significantly the size of the task structure as well as the data handle
structure.

The prototype of a codelet should now be

	void (*kernel)(void *descr[], void *arg)

The elements of the "descr" array should be pointers to the buffer container
the data description on the current memory node. For instance, if the i-th
piece of data is a "vector" data (described by the starpu_vector_interface_t
struct type), a typical way to access this data is now:

void kernel(void *descr[], void *arg)
{
	/* descr[i].vector.nx is deprecated */

	struct starpu_vector_interface_t *vector_interface = descr[i];
	unsigned nx = vector_interface->nx;
}

To somehow avoid verbosity, we provide some macros, for instance the
previous kernel could have been written as:

void kernel(void *descr[], void *arg)
{
	unsigned nx = GET_VECTOR_NX(descr[i]);
}