Fix locking around implicit dependencies
some tasks with cl==NULL are used for data dependencies, and thus appear in
last_writer/reader. they also need to be cleared from there, in case they are to
be destroyed before another writer/reader replaces them. Instead of doing it in
a callback, which is too late since we notify dependencies before that, add an
implicit_dep_handle member just for that.

Keeping the job sync_mutex held along all dependency computation (but not
always) poses deadlock issues when a job is terminating while another one
is trying to depend on it. Adding a spinlock on the completion group list
permits to decouple this: we can release it while notifying completion, and
recheck list bound after that. _starpu_add_successor_to_cg_list now returns
whether the group was completed, to keep proper atomicity. We then don't need
job_is_already_locked any more. We also don't need to check whether we have to
destroy the task any more.

_starpu_detect_implicit_data_deps_with_handle needs to submit a synchronization task, but it might be ready immediately, thus calling the callback with the job mutex held, which is not a good idea. Better make _starpu_detect_implicit_data_deps_with_handle return the task and let the caller submit it after having released the mutex

Should at last fix cg, which was the only example that really stresses keeping submitting tasks and waiting for them concurrently (with reduction in the middle, even...) :)