data interface: update the prototype of the function pack_data to be able to return a negative value for the size
Pack the data handle into a contiguous buffer at the address ptr and
set the size of the newly created buffer in count. If ptr is NULL, the
function should not copy the data in the buffer but just set count to
the size of the buffer which would have been allocated. The special value
-1 indicates the size is yet unknown.