doc: fix 'make dist' when latex environment is not found
Even if --disable-build-doc is specified to the configure script, 'make dist'
tries to build the documentation and it miserably fails when the latex
environment (especially pdflatex) is not found. This is because conditional
subdirectories must be handled manually. As specified by the automake
documentation :

"If SUBDIRS contains AC_SUBST variables, DIST_SUBDIRS will not be defined
correctly because Automake doesn't know the possible values of these
variables."'