hypervisor: fix a deadlock when using lazy resize and perf counters
This fixes a deadlock when SC_HYPERVISOR_LAZY_RESIZE is set to 1 (default
behaviour) because the mutex is already acquired in notify_post_exec_hook()
and not released before the call to _ack_resize_completed() which will again
try to lock the same mutex. This fixes the app_driven.test.c example.