openmp: fix a crash with _starpu_omp_get_task() for non-CPU workers
For the KStar heterogeneous study, we added a new 'where' clause on task
constructs. This clause can be used to tell the runtime where to execute
the tasks, the flag accepts the following values cpu, gpu and all.

To know if the current thread comes from a CPU or a GPU worker, we
hacked up omp_is_initial_device() which should return 0 for GPU workers,
and 1 for CPU workers.

That workaround is only for experimental uses and it's not going to be
part of the official specification.