Fix handle initialized flag when reading an uninitialized data
When running data_acquire(R), _starpu_data_check_initialized explicitly
creates a task that writes the data. But when submitting a task that reads
the data, it is the task which will call the cl_init function, so it is the
implicit dependency duty to note that the handle will be initialized.

(cherry picked from commit 5109603b5e0dfa749b8b396571aeb28d0e2cb75b)