Disable quick-allocation for pinned allocation
We used to optimize allocation requests by allocating immediately when it's
"just RAM". But when we are pinning it for CUDA this is actually expensive
(or worse, wait for an allocation task in the case of !CUDA_MEMCPY_PEER), so
we do rather defer it to some point when we have the time to do it (possibly
an idle worker or during a CUDA task).

(cherry picked from commit 51eda9d44cc9bc649931e1d182e6a182690a5213)