coherency: Do not collapse requests for different tasks
When we have a long list of data prefetches for various tasks, we might
not actually be able to reuse data, and thus the prefetch for the later
task will have been lost in collapsing. Also, priorities might be
different, so we'd rather let request lists keep all requests, and at
worse cancel redundant requests.