Fix missing CL_CALLBACK for various callback functions
_cl_icd_dispatch declares these function pointers with CL_CALLBACK
which expands to __stdcall, which is not the default on 32bit Windows.

Make sure to use CL_CALLBACK in all functions assigned to it, so
the right calling convention is used.

This is also motivated by clang v16 now erroring out because of this.

(cherry picked from commit 66839ac3ebee4b2b7aadc327dd4ac77475c6affe)