Fix evicting data from GPU when main ram is full of cached tiles
If we have a cached data that has the right size, we can evict to it.
This does not fix the case when tiles have different sizes, we would need to
maintain the size of the cached allocations.

(cherry picked from commit 882aba682cec925bf6bd226c210641bd80b0795d)