Fix asynchronous partitioning with data without home node
This is not that optimized for now: we just always gather only on the
gathering node, and don't benefit from free coherency on nodes which would
happen to have all the pieces. We also don't benefit from free allocation on
nodes which would happen to already have the allocation. Such optimizations
will require more thinking, as we get competition between children reusing
parent allocation, and allocation eviction. We will need to somehow make the
parent track whether it's allocation is reused by a child, but in a not too
costly way and that doesn't pose parent/child spinlock ordering issue.