This change prevents StarPU from recalibrating the machine when the user-requested configuration disables certain devices (e.g., CUDA), even if StarPU was compiled with CUDA support.
Before this commit, setting the number of CUDA devices to a positive value did not trigger a benchmarking of the bus, while setting it to 0 did.

With this modification, the bus will be calibrated the first time with all the hardware devices available on the machine, and will no longer be recalibrated for further executions regardless of the value assigned.